fix(csdlc-v2): make finish sole merge authority
