+++ .csdlc/evidence/5780/deletion-metrics.md
@@ -4,16 +4,17 @@ Baseline: `a5a16509af05af441369e78eeacfd74017bcf63b`.
 
 | Surface | Before | After | Net deleted | Reduction |
 | --- | ---: | ---: | ---: | ---: |
-| Production Rust | 28,722 | 18,516 | 10,206 | 35.53% |
-| Test Rust | 15,092 | 9,247 | 5,845 | 38.73% |
-| Combined Rust | 43,814 | 27,763 | 16,051 | 36.64% |
+| Production Rust | 28,722 | 18,194 | 10,528 | 36.65% |
+| Test Rust | 15,092 | 9,276 | 5,816 | 38.54% |
+| Combined Rust | 43,814 | 27,470 | 16,344 | 37.30% |
 | Source binaries | 23 | 21 | 2 | 8.70% |
 
 Across every changed tracked file, including lifecycle cards and proof logs,
-the implementation adds 4,379 lines and deletes 18,174, for a net deletion of
-13,795 lines.
+the implementation adds 4,674 lines and deletes 18,751, for a net deletion of
+14,077 lines.
 
-The removed surface includes the `csdlc-closeout` binary and skill, terminal
+The removed surface includes the standalone closeout and merge binaries, the
+closeout skill, terminal
 repair and reconciliation request schemas, readiness and merged-publication
 writers, legacy receipt writers, and their writer-specific tests. The retained
 compatibility boundary is read-only: all 114 tracked v0.91.8 terminal records
