+++ .csdlc/issues/5780/audit.jsonl
@@ -6,3 +6,4 @@
 {"sequence":6,"generation":2,"actor":"codex:5780-implementation","reason":"atomically record execution, validation, and implemented phase","operation":"finalize_implementation"}
 {"sequence":7,"generation":2,"actor":"codex:5780-implementation","reason":"Compiler-guided deletion reached additional exact owner, operator guidance, and obsolete proof files required to remove the competing terminal authority without leaving stale active contracts.","operation":"{\"add_protected_paths\":[\"csdlc-v2/operator/skills/csdlc-v2-finish/SKILL.md\",\"csdlc-v2/operator/skills/csdlc-v2-github/SKILL.md\",\"csdlc-v2/src/github.rs\",\"csdlc-v2/src/proof.rs\",\"csdlc-v2/tests/gate7_terminal_design_repair_5487.rs\",\"csdlc-v2/tests/gate7_terminal_sor_validation_repair_5613.rs\",\"csdlc-v2/tests/gate9.rs\"],\"operation\":\"amend_claim_scope\"}"}
 {"sequence":8,"generation":2,"actor":"codex:5780-review-remediation","reason":"Independent review found the standalone csdlc-merge binary remained a competing and unreachable terminal mutation route; remove its exact source and proof surfaces while retaining the internal finish-owned merge helper.","operation":"{\"add_protected_paths\":[\"csdlc-v2/src/bin/csdlc-merge.rs\"],\"operation\":\"amend_claim_scope\"}"}
+{"sequence":9,"generation":2,"actor":"codex:5780-review-remediation","reason":"Exact-head review found the remaining public direct-merge API and active operator guidance to be competing terminal authorities; fold merge execution into finish and guard the corrected documentation.","operation":"{\"add_protected_paths\":[\"csdlc-v2/src/bin/csdlc-finish.rs\",\"csdlc-v2/src/finish.rs\",\"csdlc-v2/src/merge.rs\",\"docs/tooling/ADL_CSDLC_GITHUB_CLIENT_BOUNDARY.md\"],\"operation\":\"amend_claim_scope\"}"}
